Stable version

+++ src/engine/generic_object.rs
@@ -9,7 +9,7 @@ pub trait GenericObject{
     fn get_texture_id(&self)->i32;
     fn get_size(&self)->(f32, f32, f32);
     fn get_texture_coordinates(&self)->((f32,f32),(f32,f32),(f32,f32),(f32,f32));
-    fn get_order(&self)->u16;
+    fn get_order(&self)->u8;
 }
 
 impl <F: ?Sized> GenericObject for Box<F>
@@ -39,7 +39,7 @@ impl <F: ?Sized> GenericObject for Box<F>
     fn get_texture_coordinates(&self)->((f32,f32),(f32,f32),(f32,f32),(f32,f32)){
         (**self).get_texture_coordinates()
     }
-    fn get_order(&self)->u16{
+    fn get_order(&self)->u8{
         (**self).get_order()
     }
